refactor: move game files to components package
