--- internal/tui/input.go
@@ -1,2 +1,0 @@
-package tui
-
